refactor: move websocket dialer and connection to own pkg
