+++ pkg/client/events_query/client.go
@@ -12,6 +12,7 @@ import (
 	"go.uber.org/multierr"
 
 	"pocket/pkg/client"
+	"pocket/pkg/client/events_query/websocket"
 	"pocket/pkg/either"
 	"pocket/pkg/observable"
 	"pocket/pkg/observable/channel"
@@ -66,7 +67,7 @@ func NewEventsQueryClient(cometWebsocketURL string, opts ...client.EventsQueryCl
 
 	if evtClient.dialer == nil {
 		// default to using the websocket dialer
-		evtClient.dialer = NewWebsocketDialer()
+		evtClient.dialer = websocket.NewWebsocketDialer()
 	}
 
 	return evtClient
@@ -136,6 +137,7 @@ func (eqc *eventsQueryClient) newEventsBytesAndConn(
 	ctx context.Context,
 	query string,
 ) (*eventsBytesAndConn, error) {
+	// Get a connection for the query.
 	conn, err := eqc.openEventsBytesAndConn(ctx, query)
 	if err != nil {
 		return nil, err
@@ -159,8 +161,7 @@ func (eqc *eventsQueryClient) openEventsBytesAndConn(
 	ctx context.Context,
 	query string,
 ) (client.Connection, error) {
-	// If no event subscription exists for the given query, create a new one.
-	// Generate a new unique request ID.
+	// Get a request for subscribing to events matching the given query.
 	req, err := eqc.eventSubscriptionRequest(query)
 	if err != nil {
 		return nil, err
@@ -243,17 +244,14 @@ func (eqc *eventsQueryClient) goUnsubscribeOnDone(
 }
 
 // eventSubscriptionRequest returns a JSON-RPC request for subscribing to events
-// matching the given query.
+// matching the given query. The request is serialized as JSON to a byte slice.
 // (see: https://github.com/cometbft/cometbft/blob/main/rpc/client/http/http.go#L110)
 // (see: https://github.com/cosmos/cosmos-sdk/blob/main/client/rpc/tx.go#L114)
 func (eqc *eventsQueryClient) eventSubscriptionRequest(query string) ([]byte, error) {
-	// Generate a new unique request ID, size and keyspace space are arbitrary.
-	requestId := randRequestId()
-
 	requestJson := map[string]any{
 		"jsonrpc": "2.0",
 		"method":  "subscribe",
-		"id":      requestId,
+		"id":      randRequestId(),
 		"params": map[string]interface{}{
 			"query": query,
 		},
@@ -266,7 +264,10 @@ func (eqc *eventsQueryClient) eventSubscriptionRequest(query string) ([]byte, er
 }
 
 // randRequestId returns a random 8 byte, base64 request ID which is intended
-// for in JSON-RPC requests to uniquely identify distinct RPC requests. These IDs
+// for in JSON-RPC requests to uniquely identify distinct RPC requests.
+// These request IDs only need to be unique to the extent that they are useful
+// to this client for identifying distinct RPC requests.
+// These IDs
 // are expected to be unique (per request). Its size and keyspace are arbitrary.
 func randRequestId() string {
 	requestIdBz := make([]byte, 8) // 8 bytes = 64 bits = uint64
