+++ pkg/client/events_query/errors.go
@@ -6,6 +6,6 @@ var (
 	ErrDial       = errorsmod.Register(codespace, 1, "dialing for connection failed")
 	ErrConnClosed = errorsmod.Register(codespace, 2, "connection closed")
 	ErrSubscribe  = errorsmod.Register(codespace, 3, "failed to subscribe to events")
-	ErrReceive    = errorsmod.Register(codespace, 4, "failed to receive event")
-	codespace     = "events_query_client"
+
+	codespace = "events_query_client"
 )
