+++ pkg/client/events_query/websocket/dialer.go
@@ -1,4 +1,4 @@
-package eventsquery
+package websocket
 
 import (
 	"context"
